Implement writing Metapost code to generate scorecard
This commit adds the basis for writing Metapost code to generate the
final scorecard. At this point, everything but the play-but-play and
individual inning totals are being outputted.

Additionally, a compilation file has been added to the examples, to
provide a way to compile the metapost files into a PDF in the meantime.