Fix issue with non-numbered plays on basepaths causing issues.
On the 2018-04-03 BOS @ MIA game, one of the plays is a runner
advancing on the throw. This is labeled as "T" from 2B to 3B, which
caused issues with the WayBase code, since it expects a number and a
play, rather than just a play, which caused a crash.

This issue has been fixed by checking whether there is a jersey
number to print or not.