Fix issue with defensive substitutions in away team
Issue:
Defensive substitutions on the away team were not being printed in
the correct inning.

Cause:
The code for defensive substitutions was not taking into account that
for the away team, the replacement line has to be drawn on the
pinch-runner side of the inning, rather than the pinch-hitter's side.

Fix:
If the defensive substitution occurs for the away team, the X start
is moved one at-bat width to match the correct inning.