Make the Metapost template dynamic depending on number of innings
This commit adds the ability for the Metapost template to resize
depending on the number of innings played.
By default, 15 innings should be used.

Additionally, the out number label has been moved from the bottom
right corner of the diamond, to the top right corner, to avoid
clashing with text on the 1B basepath.