Allow for thrown out calls to list the current base
In some scenarios, MLB will list the out base of a runner getting
thrown out as the current base, such as in pickoffs or in double plays
where the runner doesn't return to base on time after an out on air.

The thrown_out method has been enhanced to allow for these scenarios.